feat: Specify scan job timeout for Polaris scanner
