+++ pkg/cmd/find_vulnerabilities.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"context"
-	"time"
 
 	"github.com/aquasecurity/starboard/pkg/find/vulnerabilities/crd"
 	"github.com/aquasecurity/starboard/pkg/find/vulnerabilities/trivy"
@@ -10,10 +9,6 @@ import (
 	"github.com/spf13/cobra"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	"k8s.io/client-go/kubernetes"
-)
-
-const (
-	scanJobTimeoutFlagName = "scan-job-timeout"
 )
 
 func GetVulnerabilitiesCmd(cf *genericclioptions.ConfigFlags) *cobra.Command {
@@ -87,17 +82,7 @@ NAME is the name of a particular Kubernetes workload.
 		},
 	}
 
-	cmd.Flags().Duration(scanJobTimeoutFlagName, time.Duration(0),
-		"The length of time to wait before giving up on a scan job. Non-zero values should contain a"+
-			" corresponding time unit (e.g. 1s, 2m, 3h). A value of zero means don't timeout the scan job.")
+	registerScannerOpts(cmd)
 
 	return cmd
-}
-
-func getScannerOpts(cmd *cobra.Command) (opts trivy.ScannerOpts, err error) {
-	opts.ScanJobTimeout, err = cmd.Flags().GetDuration(scanJobTimeoutFlagName)
-	if err != nil {
-		return
-	}
-	return
 }
