+++ pkg/scanners/base_test.go
@@ -1,0 +1,1 @@
+package scanners
